STEP_2. Integration of nested levels in comments
-new fields in Comment - level, parentId, Image
-new fields in DB
-roles written in DB

+++ src/main/java/com/github/roveraven/service/CommentService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface CommentService {
     Comment save(Comment comment);
-    Optional<Comment> findComment(Integer CommentId);
+    Optional<Comment> findComment(Long CommentId);
     List<Comment> findByUser(Long userId);
     List<Comment> findAll();
